moved camera logic into its own file
