+++ src/game/mod.rs
@@ -8,6 +8,7 @@ pub mod assets;
 pub mod audio;
 mod movement;
 pub mod spawn;
+mod camera;
 
 pub(super) fn plugin(app: &mut App) {
     app.add_plugins((
@@ -15,5 +16,6 @@ pub(super) fn plugin(app: &mut App) {
         audio::plugin,
         movement::plugin,
         spawn::plugin,
+        camera::plugin,
     ));
 }
